foobar dependency on mount unit works.
The key is that require should not be specified, otherwise, if mount
fails, foobar won't be brough up.

Remove this option will bring up foobar each time mount command is
issued, no matter if it succeeds. This is what we want.

Further, noticing some other services requires `require` option. For
example, samba should no be started if mount fails. Unlike foobar, which
is required to start anyway.

In short, foobar just needs to start after mount, it is an ORDERING issue,
not really a dependency.