Change uuid file and volume folder definition.
uuid is a normal file. it exists at any case, even if there is no uuid in
mount unit file. In this case, the file is simply touched.

volume is a folder or a symlink to folder. if uuid is set AND there is
at least one disk in chassis ( the volume is created ), a symlink is
created, following the correspoding volume folder. Otherwise, an empty
folder is created.